+++ setup.php
@@ -6,7 +6,7 @@
  * @license GPL-2.0-or-later
  */
 
-define('PLUGIN_MAILBLAST_VERSION',  '1.0.0');
+define('PLUGIN_MAILBLAST_VERSION',  '1.0.1');
 define('PLUGIN_MAILBLAST_MIN_GLPI', '11.0.0');
 define('PLUGIN_MAILBLAST_MAX_GLPI', '11.99.99');